[vxlan_ecmp] fix some issues after moving to python3 (#8658)

+++ tests/vxlan/vxlan_ecmp_utils.py
@@ -578,8 +578,8 @@ class Ecmp_Utils(object):
             netid  : The first octet.
         '''
         third_octet = self.Address_Count % 255
-        second_octet = (self.Address_Count / 255) % 255
-        first_octet = netid + (self.Address_Count / 65025)
+        second_octet = int(self.Address_Count / 255) % 255
+        first_octet = netid + int(self.Address_Count / 65025)
         self.Address_Count = self.Address_Count + 1
         if af == 'v4':
             return "{}.{}.{}.{}".format(
